requirements.txt
tiktoken
openai
langchain
cassio
datasets
python-dotenv
PyPDF2
streamlit